chore: moving koa to instrumentation
