+++ examples/koa/package.json
@@ -35,7 +35,7 @@
     "@opentelemetry/api": "^0.15.0",
     "@opentelemetry/exporter-jaeger": "^0.15.0",
     "@opentelemetry/exporter-zipkin": "^0.15.0",
-    "@opentelemetry/koa-instrumentation": "^0.13.0",
+    "@opentelemetry/instrumentation-koa": "^0.13.0",
     "@opentelemetry/node": "^0.15.0",
     "@opentelemetry/plugin-http": "^0.15.0",
     "@opentelemetry/tracing": "^0.15.0",
