+++ package.json
@@ -3,8 +3,8 @@
   "version": "0.13.1",
   "description": "This is a repository for OpenTelemetry JavaScript contributions.",
   "repository": {
-    "type" : "git",
-    "url" : "https://github.com/open-telemetry/opentelemetry-js-contrib.git"
+    "type": "git",
+    "url": "https://github.com/open-telemetry/opentelemetry-js-contrib.git"
   },
   "publishConfig": {
     "access": "public"
