+++ plugins/node/opentelemetry-instrumentation-koa/.npmignore
